Moving `GitCommit` to a separate package and changing name to `Commit`
